make virus experiments work

+++ virus_trees.py
@@ -30,9 +30,8 @@ def run_inference(msa_path, model, prefix, args = ""):
 
 def run_root_digger(msa_path, tree_path):
     command = rd_path
-    command += " --msa /" + msa_path
-    command += " --tree /" + tree_path
-    command += "  --exhaustive "
+    command += " --msa " + msa_path
+    command += " --tree " + tree_path
     print(command)
     os.system(command)
     suffixes = [".ckp" , "lwr.tree", ".rooted.tree"]
@@ -40,7 +39,7 @@ def run_root_digger(msa_path, tree_path):
 
 
 raxmlng_path = "./bin/raxml-ng"
-rd_path = "docker run -v $(pwd)/data/:/data/ my_rd:latest"
+rd_path = "./bin/rd"
 
 for virus_type in ["rsv"]:
     base_dir = os.path.join("data/virus", virus_type)
@@ -64,8 +63,8 @@ for virus_type in ["rsv"]:
         #best_tree_path = get_best_tree_path(prefix)
         unrooted_tree_path = os.path.join(unrooted_trees_dir, msa_name_x + ".unrooted.tree")
         #shutil.copyfile(best_tree_path, unrooted_tree_path)
-        run_root_digger(os.path.join(msa_dir, msa_name), unrooted_tree_path)
-        shutil.copyfile(unrooted_tree_path + ".rooted.tree", rooted_tree_path)
+        #run_root_digger(os.path.join(msa_dir, msa_name), unrooted_tree_path)
+        #shutil.copyfile(unrooted_tree_path + ".rooted.tree", rooted_tree_path)
         suffixes = [".ckp" , ".lwr.tree", ".rooted.tree"]
         for suffix in suffixes:
             src = unrooted_tree_path + suffix
@@ -73,4 +72,5 @@ for virus_type in ["rsv"]:
             if not os.path.isdir(dst_dir):
                 os.makedirs(dst_dir)
             dst = os.path.join(dst_dir, "rd" + suffix)
-            shutil.move(src, dst)
+            if os.path.isfile(src):
+                shutil.move(src, dst)
